+++ SpreadsheetUtility.UI.Web/Components/Pages/GanttGeneratorFromPaste.razor
@@ -47,8 +47,8 @@
     <QuickGrid Items="@taskOutput">
         <PropertyColumn Property="@(p => p.Id)" Sortable="true" />
         <PropertyColumn Property="@(p => p.Name)" Sortable="true" />
-        <PropertyColumn Property="@(p => p.StartDate)" Sortable="true" />
-        <PropertyColumn Property="@(p => p.EndDate)" Sortable="true" />
+        <PropertyColumn Property="@(p => p.StartDate.ToString("dd-MM-yyyy"))" Sortable="true" />
+        <PropertyColumn Property="@(p => p.EndDate.ToString("dd-MM-yyyy"))" Sortable="true" />
         <PropertyColumn Property="@(p => p.EstimatedEffortHours)" Sortable="true" />
         <PropertyColumn Property="@(p => p.Dependencies)" Sortable="true" />
     </QuickGrid>
@@ -61,7 +61,7 @@
         <PropertyColumn Property="@(p => p.DailyWorkHours)" Sortable="true" />
         <PropertyColumn Property="@(p => p.AllocatedHours)" Sortable="true" />
         <PropertyColumn Property="@(p => p.SlackHours)" Sortable="true" />
-        <PropertyColumn Property="@(p => p.NextAvailableDateForTasks)" Sortable="true" />        
+        <PropertyColumn Property="@(p => p.NextAvailableDateForTasks.ToString("dd-MM-yyyy"))" Sortable="true" />        
         <PropertyColumn Property="@(p => p.VacationPeriodsSerialized)" Sortable="true" />
     </QuickGrid>
 </div>
