+++ SpreadsheetUtility/SpreadsheetUtility.Library/Processors/SpreadsheetGenerator/SpreasheetGeneratorDoubleEntry.cs
