+++ SpreadsheetUtility/SpreadsheetUtility.Library/Services/GanttService/GanttService.cs
@@ -7,15 +7,7 @@ using System.Threading.Tasks;
 
 namespace SpreadsheetUtility.Services
 {
-    public interface IGanttService
-    {
-        string ProcessExcelDataTasksFromFile(string taskFilePath, string teamFilePath);
-        string ProcessExcelDataProjectsFromFile(string taskFilePath, string teamFilePath);
-        List<GanttTask> LoadTasksFromDtos(List<TaskDto> taskDtos);
-        List<DeveloperAvailability> LoadDeveloperAvailabilityFromDtos(List<DeveloperDto> taskDtos);
-        List<GanttTask> AssignProjectsFromDtos(List<TaskDto> taskDtos, List<DeveloperDto> developerDtos);
-        GanttChartAllocation CalculateGanttChartAllocationFromDtos(List<TaskDto> taskDtos, List<DeveloperDto> developerDtos);
-    }
+    
     public class GanttService : IGanttService
     {
         IGanttChartProcessor _ganttChartProcessor;
